+++ guardrails_genie/guardrails/privilege_escalation/__init__.py
@@ -1,4 +1,4 @@
-from guardrails_genie.guardrails.privilege_escalation.priv_esc_guardrail import (
+from guardrails_genie.guardrails.privilege_escalation.priv_esc_guardrails import (
     PrivilegeEscalationGuardrail,
 )
